Rename the AdminMenuTooltip folder to AdminScreenTooltip.
